make notification messages simpler
