+++ WalletWasabi.Fluent/Converters/MoneyConverters.cs
@@ -1,6 +1,6 @@
-using System;
 using Avalonia.Data.Converters;
 using NBitcoin;
+using WalletWasabi.Fluent.Helpers;
 
 namespace WalletWasabi.Fluent.Converters
 {
@@ -14,30 +14,6 @@ namespace WalletWasabi.Fluent.Converters
 			});
 
 		public static readonly IValueConverter ToFormattedString =
-			new FuncValueConverter<Money?, string>(money =>
-			{
-				if (money is null)
-				{
-					return "";
-				}
-
-				const int WholeGroupSize = 3;
-
-				var moneyString = money.ToString();
-
-				moneyString = moneyString.Insert(moneyString.Length - 4, " ");
-
-				var startIndex = moneyString.IndexOf(".", StringComparison.Ordinal) - WholeGroupSize;
-
-				if (startIndex > 0)
-				{
-					for (var i = startIndex; i > 0; i -= WholeGroupSize)
-					{
-						moneyString = moneyString.Insert(i, " ");
-					}
-				}
-
-				return moneyString;
-			});
+			new FuncValueConverter<Money?, string>(money => money is null ? "" : money.ToFormattedString());
 	}
 }
